@@ -5,12 +5,12 @@
   "repository": "github/paste-markdown",
   "main": "dist/index.umd.js",
   "module": "dist/index.esm.js",
-  "types": "index.d.ts",
+  "types": "dist/index.d.ts",
   "scripts": {
     "clean": "rm -rf dist",
     "lint": "github-lint",
     "prebuild": "npm run clean && npm run lint && mkdir dist",
-    "build": "rollup -c && cp src/index.js.flow dist/index.umd.js.flow && cp src/index.js.flow dist/index.esm.js.flow",
+    "build": "rollup -c",
     "pretest": "npm run build",
     "test": "karma start karma.config.js",
     "prepublishOnly": "npm run build",
@@ -22,24 +22,22 @@
   ],
   "license": "MIT",
   "files": [
-    "dist",
-    "index.d.ts"
+    "dist/index.d.ts",
+    "dist/index.esm.js",
+    "dist/index.umd.js"
   ],
   "devDependencies": {
-    "@babel/core": "^7.7.0",
-    "babel-eslint": "^10.0.3",
-    "babel-preset-github": "^3.2.1",
     "chai": "^4.2.0",
-    "eslint": "^6.6.0",
-    "eslint-plugin-github": "^3.2.1",
-    "flow-bin": "^0.111.1",
+    "eslint": "^6.8.0",
+    "eslint-plugin-github": "^3.4.0",
     "karma": "^4.4.1",
     "karma-chai": "^0.1.0",
     "karma-chrome-launcher": "^3.1.0",
     "karma-mocha": "^1.3.0",
     "karma-mocha-reporter": "^2.2.5",
     "mocha": "^7.0.1",
-    "rollup": "^1.26.3",
-    "rollup-plugin-babel": "^4.3.3"
+    "rollup": "^1.31.1",
+    "rollup-plugin-typescript2": "^0.26.0",
+    "typescript": "^3.7.5"
   }
 }
