@@ -1,8 +1,6 @@
-/* @flow strict */
-
 export function insertText(textarea: HTMLInputElement | HTMLTextAreaElement, text: string): void {
-  const beginning = textarea.value.substring(0, textarea.selectionStart)
-  const remaining = textarea.value.substring(textarea.selectionEnd)
+  const beginning = textarea.value.substring(0, textarea.selectionStart || 0)
+  const remaining = textarea.value.substring(textarea.selectionEnd || 0)
 
   const newline = beginning.length === 0 || beginning.match(/\n$/) ? '' : '\n'
   const textBeforeCursor = beginning + newline + text
